+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firElement.kt
@@ -1,0 +1,8 @@
+package com.javiersc.kotlin.compiler.extensions.fir
+
+import org.jetbrains.kotlin.fir.FirElement
+import org.jetbrains.kotlin.fir.render
+
+public fun List<FirElement?>.render(): String = buildString {
+    for (firElement in this@render) appendLine(firElement?.render())
+}